pyproject.toml
[tool.poetry]
name = "pyspark-utils"
version = "0.1.0"
description = "PySpark reusable scripts for ETL and ML"
authors = ["Avinash Kanumuru <avinash.eee@gmail.com>"]
license = "BSD"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
pyspark = "3.5.2"
boto3 = "^1.34.129"
pyyaml = "^6.0.1"
loguru = "^0.7.2"
findspark = "^2.0.1"

[tool.poetry.group.dev.dependencies]
pdoc = "^14.5.0"
pytest = "^8.2.2"
pandas = {extras = ["pyarrow"], version = "^2.2.2"}
numpy = "<2"
ipykernel = "^6.29.4"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
